feat(receipts): add multi-select product filtering for receipts
- Enhance product filter to support multiple comma-separated product names
- Add autocomplete UI component for product selection with tag display
- Update Receipt.number_receipt to BigIntegerField for larger values
- Improve filter UI layout with better styling and user experience
- Add safety checks for window.tokens in JavaScript
- Simplify script loading logic for receipts pages
- Replace receipt_filter.js with new receipts-filter.js implementation

+++ hasta_la_vista_money/receipts/models.py
@@ -396,7 +396,7 @@ class Receipt(models.Model):
     """
 
     receipt_date = models.DateTimeField()
-    number_receipt = models.IntegerField(default=None, null=True)
+    number_receipt = models.BigIntegerField(default=None, null=True)
     nds10 = models.DecimalField(
         blank=True,
         max_digits=constants.SIXTY,
